Enable clang-diagnostic-overloaded-virtual
28 sites in 18 files, measured over all 840 translation units in
cmake-build-strict-ra-clang-22 and deduplicated by site. The sample in
docs/CLANG_TIDY_PRIORITIES.md predicted 10.

Six were live bugs. A shadowed virtual is not an override, so the derived
version never runs through a base pointer and is dead:

- VesselClass::Take_Damage took 'int forced' where ObjectClass takes 'bool', so
  ship death announcements, explosion animations and screen shake were
  unreachable.
- TD's TurretClass::Ok_To_Move is non-const because it sets SecondaryFacing;
  DriveClass::Ok_To_Move is const. Turret-locked vehicles therefore never
  waited for the turret to come round before moving.
- RA's ObjectClass::Overlap_List gained a 'bool redraw' parameter that
  AnimClass and BulletClass never picked up, so their refresh regions were
  unused. TD is the mirror image: Occupy_List gained 'bool placement' there,
  and TD's AnimClass and BulletClass were left behind instead.
- ListClass::Get_Item took size_t in RA while every override took int, so
  CheckListClass and EListClass silently stopped overriding it. Moved the base
  to int, matching TD's copy of the same class and the signed-index rule in
  CLAUDE.md; the clamp now bounds both ends rather than relying on a negative
  index wrapping to a huge size_t.
- Four WinNullModemClass stubs still took char* after WinModemClass was
  constified, so a null-modem link fell through to the real modem code.

Each of those is fixed by matching the base signature, which means they now
actually run. That is a behaviour change in both games, and it is the point of
the check.

ColorListClass::Add_Item was a latent one: the defaulted colour argument hid
both base overloads, and Colors is a parallel vector that Draw_Entry indexes
unconditionally, so an add through a ListClass& would have desynchronised the
two. The base signatures are now real overrides that forward with the default
colour, so no existing call changes.

BuildingClass::Remap_Table is deleted rather than revived. It was already dead
by the same shadowing, and its comment -- that the base "presumes the object is
a unit" -- no longer holds: TechnoClass::Remap_Table goes through
Techno_Type_Class() and adds an IsRemappable check the building version lacks.
Reviving it would regress non-remappable buildings to the house colour;
deleting it changes nothing.

The rest were deliberate overloads, not mistakes. TFixedIHeapClass's typed
ID/Logical_ID and NullModemConnClass::Init(HANDLE) are additions to the base
API, so the hidden versions are re-exposed with a using-declaration.
IPXGlobalConnClass's three routines keep the base hidden and carry a NOLINT
with the reason: a global packet has a GlobalHeaderType with a magic number and
a destination address, and the inherited point-to-point versions write neither,
so exposing them would let a caller put a malformed packet on the wire. Those
six NOLINTs are the only suppressions.

Measures 0 sites tree-wide afterwards over the same 840 units, with
cppcoreguidelines-virtual-class-destructor and clang-diagnostic-non-virtual-dtor
unchanged at 32 and 27, and no hard error: lines. Both games build and the 84
tests pass.

+++ src/ra/list.h
@@ -81,7 +81,7 @@ class ListClass : public ControlClass {
   virtual int Current_Index() const;
   virtual const char* Current_Item() const;
   int Draw_Me(bool forced) override;
-  virtual const char* Get_Item(size_t index) const;
+  virtual const char* Get_Item(int index) const;
   virtual int Step_Selected_Index(int forward);
   void Flag_To_Redraw() override;
